Handlers.py errors corrected.
