+++ server.py
@@ -1,34 +1,24 @@
-import datetime
 import os
 
 from flask import Flask
-from flask import render_template
+from flask import Blueprint, render_template
+from handlers import site
 
 
-app = Flask(__name__)
-
-@app.route('/')
-def home_page():
-    return render_template('home/index.html')
-
-@app.route('/restaurant')
-def restaurant_home_page():
-    return render_template('restaurant/index.html')
-
-@app.route('/restaurant/12') #Change me with model [ID]
-def restaurant_show_page():
-    return render_template('restaurant/show.html')
-
-@app.route('/register')
-def register_home_page():
-    return render_template('register/index.html')
+def create_app():
+    app = Flask(__name__)
+    app.register_blueprint(site)
+    return app
 
 
-
-if __name__ == '__main__':
+def main():
     VCAP_APP_PORT = os.getenv('VCAP_APP_PORT')
     if VCAP_APP_PORT is not None:
         port, debug = int(VCAP_APP_PORT), False
     else:
         port, debug = 5000, True
-    app.run(host='0.0.0.0', port=port, debug=debug)
+    app = create_app()
+    app.run(host='0.0.0.0', port=port, debug=True)
+
+if __name__ == '__main__':
+    main()
