+++ templates/restaurant/index.html
@@ -24,7 +24,7 @@
           <tbody>
             <tr>
               <!-- Change me with model and use for loop for showing all item inside the database. -->
-              <td><a href="{{ url_for('restaurant_show_page') }}">KemalPasa Koftecisi</a></td>
+              <td><a href="{{ url_for('site.restaurant_show_page') }}">KemalPasa Koftecisi</a></td>
               <td>Mahmut Hava</td>
               <td>Taksim 42/5 Beyoglu/Istanbul</td>
               <td>4/5</td>
